Finished implementing the Dictionary Builder. Replaced dictionaries with the new ones.

+++ src/enums/DecimalPrecision.java
@@ -25,6 +25,6 @@ public enum DecimalPrecision {
     }
     
     public double round(double value) {
-        return Math.round((double) value / precision) * precision;
+        return Math.round(value / precision) * precision;
     }
 }
